+++ test-project-paths-aliases/src/used_components.ts
@@ -1,0 +1,3 @@
+export function used_component() {
+  return true;
+}