+++ test-project-paths-aliases/utils/functions/used_functions2.ts
@@ -1,0 +1,3 @@
+export function used_function2() {
+  return true;
+}