+++ test-project-paths-aliases/utils/functions/used_functions3.ts
@@ -1,0 +1,3 @@
+export function used_function3() {
+  return true;
+}